pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "keynav"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyaml>=25.7.0",
    "readchar>=4.2.1",
]

[project.scripts]
keynav = "keynav.cli:main"
